remove useless extensions + divided big files in Cells
