+++ Source/TonSwift/Cells/BocMagic.swift
@@ -1,0 +1,12 @@
+//
+//  BocMagic.swift
+//  
+//
+//  Created by Kirill Kirilenko on 20/07/2023.
+//
+
+enum BocMagic: UInt32 {
+    case V1 = 0x68ff65f3
+    case V2 = 0xacc3a728
+    case V3 = 0xb5ee9c72
+}
