+++ Source/TonSwift/Cells/Cell/CellMetrics.swift
@@ -1,0 +1,14 @@
+//
+//  CellMetrics.swift
+//  
+//
+//  Created by Kirill Kirilenko on 20/07/2023.
+//
+
+/// Metrics of a cell or a builder
+public struct CellMetrics {
+    /// Number of bits in the cell
+    var bitsCount: Int
+    /// Number of refs in the cell
+    var refsCount: Int
+}
